+++ esbuild.config.js
@@ -6,12 +6,12 @@ esbuild
     entryPoints: [
       './src/index.ts',
       './src/create-preview.ts',
-      './src/update-img-paths.ts',
+      './src/ipfs-pack.ts',
     ],
     bundle: true,
     minify: true,
     outdir: 'build',
     platform: 'node',
-    external: ['canvas', 'svgo', 'cosmiconfig'],
+    external: ['canvas', 'svgo', 'cosmiconfig', 'ipfs-car'],
   })
   .catch(() => process.exit(1));